Derive lammps_pace and lammps_snap from a shared lammps_base class
